server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
pydantic>=2.0.0
